putting existing code into Cogs and the right folders
